requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langfuse>=2.0.0
streamlit>=1.38.0
python-dotenv>=1.0.0
pydantic>=2.0.0
